Update build tools

+++ build/tools/flash_kernel.sh
@@ -18,6 +18,8 @@
 selinx=$(cat /tmp/aroma/sel.prop | cut -d '=' -f2)
 qc=$(cat /tmp/aroma/crate.prop | cut -d '=' -f2)
 therm=$(cat /tmp/aroma/thermal.prop | cut -d '=' -f2)
+net=$(cat /tmp/aroma/netmode.prop | cut -d '=' -f2)
+jk=$(cat /tmp/aroma/jack.prop | cut -d '=' -f2)
 #force permissive
 selinx=3
 zim=/tmp/Image1
@@ -32,6 +34,14 @@ cmd=$cmd" androidboot.selinux=enforcing"
 elif [ $selinx -eq 3 ]; then
 cmd=$cmd" androidboot.selinux=permissive"
 fi
+if [ $net -eq 1 ]; then
+cmd=$cmd" android.gdxnetlink=old"
+elif [ $net -eq 2 ]; then
+cmd=$cmd" android.gdxnetlink=los"
+fi
+if [ $jk -eq 2 ]; then
+cmd=$cmd" android.audiojackmode=stock"
+fi
 if [ $therm -eq 1 ]; then
 echo "Using old thermal engine"
 cp -rf /tmp/old-thermal/* /system/vendor/
@@ -40,7 +50,9 @@ chmod 0644 /system/vendor/lib/libthermalclient.so
 chmod 0644 /system/vendor/lib64/libthermalclient.so
 chmod 0644 /system/vendor/lib64/libthermalioctl.so
 fi
+if ! [ -f /system/etc/radon.sh ]; then
 cp /tmp/radon.sh /system/etc/radon.sh
+fi
 chmod 644 /system/etc/radon.sh
 cp -f /tmp/cpio /sbin/cpio
 cd /tmp/
